+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
 - `f-badge` 组件 type 新增 default 配置项
+- `f-badge` 组件 `color` 配置项改为 `background`
+- `f-badge` 组件 `font-color` 配置项改为 `color`
 
 ## 0.20.3 (2023-01-31)